rename BuenaToken to BuenoToken
