+++ packages/subgraph/package.json
@@ -1,13 +1,13 @@
 {
-  "name": "buenatoken",
+  "name": "buenotoken",
   "license": "UNLICENSED",
   "scripts": {
     "codegen": "graph codegen",
     "build": "graph build",
-    "deploy": "graph deploy --node https://api.studio.thegraph.com/deploy/ buenatoken",
-    "create-local": "graph create --node http://localhost:8020/ buenatoken",
-    "remove-local": "graph remove --node http://localhost:8020/ buenatoken",
-    "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 buenatoken",
+    "deploy": "graph deploy --node https://api.studio.thegraph.com/deploy/ buenotoken",
+    "create-local": "graph create --node http://localhost:8020/ buenotoken",
+    "remove-local": "graph remove --node http://localhost:8020/ buenotoken",
+    "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 buenotoken",
     "test": "graph test"
   },
   "dependencies": {
